refactor: Split logging in example in seperate file…
…demonstrating typical usage in a project.
