+++ examples/log-complete/ecomponent.hpp
@@ -1,0 +1,15 @@
+
+// Copyright Henrik Steffen Gaßmann 2023
+//
+// Distributed under the Boost Software License, Version 1.0.
+//         (See accompanying file LICENSE or copy at
+//           https://www.boost.org/LICENSE_1_0.txt)
+
+#pragma once
+
+namespace dlog_ex
+{
+
+void do_output();
+
+}
